달력을 선택적으로 그리는 중간 컴포넌트 생성
