+++ src/component/calendar.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import {dateToString} from '../js/common';
-import CalendarBody from './body';
+import CalendarGenerator from '../component/CalendarGenerator';
 
 const Calendar = props => {
     const arrPrpDate =  /^(\d{4})[.|-|/]*(\d{2})[.|-|/]*(\d{2})$/.exec( dateToString(props.date) );    
@@ -45,7 +45,7 @@ const Calendar = props => {
         <>
             {props.children}<br/>
             <a href={href} onClick={ ()=> prevMonth() }>&lt;</a>&nbsp;&nbsp; {calObj.year} / {calObj.month} &nbsp;&nbsp;<a href={href} onClick={ ()=> nextMonth() }>&gt;</a>
-            <CalendarBody year={calObj.year} month={calObj.month} day={calObj.day} date={calObj.thisDate}></CalendarBody>
+            <CalendarGenerator year={calObj.year} month={calObj.month} day={calObj.day} date={calObj.thisDate}></CalendarGenerator>
         </>
     )
 }
