+++ src/js/common.js
@@ -18,4 +18,17 @@ export const dateToString = (str) => {
     return `${date.getFullYear()}${mon}${day}`
 }
 
+export const genDateList = prp => {
+    const firstDay = new Date( `${prp.year}/${prp.month}/01` ).getDay();    
+    const lastDay = getLastDate(prp.date);        
+    return  Array(firstDay).fill(null).concat( Array(lastDay).fill(null).map( 
+                (item,idx) => {
+                    return { 
+                        fullDate : `${prp.year}${prp.month}${ idx+1 >= 10 ? idx+1 : '0'+(idx+1) }`
+                        ,date : `${ idx+1 >= 10 ? idx+1 : '0'+(idx+1) }`
+                    }
+                }
+            )
+        )        
+}
